requirements.txt
absl-py==1.3.0
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==1.1.0
altair==4.2.0
aniso8601==9.0.1
annotated-types==0.5.0
antlr4-python3-runtime==4.8
anyio==3.7.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.1.0
autofaiss==2.15.6
backcall==0.2.0
basicsr==1.4.2
blinker==1.6.2
braceexpand==0.1.7
cachetools==5.2.0
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.3
clip-anytorch==2.5.2
clip-retrieval==2.36.1
cmake==3.27.0
comm==0.1.4
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.35
dataclasses==0.6
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
diffusers==0.21.4
dlib==19.24.2
docker-pycreds==0.4.0
easydict==1.10
einops==0.4.1
embedding-reader==1.5.0
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
ExifRead-nocycle==3.0.1
face-alignment==1.4.1
facexlib==0.2.5
faiss-cpu==1.7.4
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.8.0
filterpy==1.4.5
fire==0.4.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.10
GitPython==3.1.29
google-auth==2.22.0
google-auth-oauthlib==1.0.0
gradio==3.46.1
gradio_client==0.5.3
grpcio==1.50.0
h11==0.14.0
h5py==3.8.0
headpose==1.1.8
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
idna==3.4
imageio==2.14.1
imageio-ffmpeg==0.4.7
img2dataset==1.41.0
importlib-metadata==6.8.0
importlib-resources==5.10.0
imutils==0.5.4
ipykernel==6.25.0
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
kornia==0.6.0
lit==16.0.6
llvmlite==0.39.1
lmdb==1.4.1
lpips==0.1.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.2
multilingual-clip==1.0.10
nest-asyncio==1.5.7
networkx==2.8.8
ninja==1.11.1
nltk==3.8.1
numba==0.56.4
numexpr==2.8.5
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.1.1
open-clip-torch==2.19.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.1.78
orjson==3.9.7
packaging==21.3
pandas==1.5.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pudb==2019.2
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.7
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.0
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
PyQt5==5.13.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
PyQt6-Qt6==6.5.2
PyQt6-sip==13.5.2
pyrsistent==0.19.2
python-dateutil==2.8.2
python-multipart==0.0.6
pytorch-lightning==1.5.9
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qdarkgraystyle==1.0.2
QDarkStyle==3.1
QtPy==2.4.0
qudida==0.0.4
regex==2022.10.31
requests==2.31.0
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
sacremoses==0.0.53
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
semantic-version==2.10.0
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.21.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
streamlit==1.15.1
supervision==0.15.0
sympy==1.12
tb-nightly==2.15.0a20230817
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
termcolor==2.3.0
test-tube==0.7.5
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.5.4
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchvision==0.15.2
tornado==6.3.2
tqdm==4.64.1
traitlets==5.9.0
transformers==4.18.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.13
urwid==2.1.2
uvicorn==0.23.2
validators==0.20.0
wandb==0.12.21
watchdog==2.1.9
wcwidth==0.2.6
webdataset==0.2.48
websockets==11.0.3
Werkzeug==2.3.3
wrapt==1.15.0
yacs==0.1.8
yapf==0.40.1
yarl==1.8.1
zipp==3.10.0
